CoreBE/api/requirement.txt
uvicorn
fastapi
redis
minio
python-dotenv